Simplify management of active message scopes by tracking them per-session-thread
This assumes that if a thread receives a message from the same session, any previous message scope
associated with that thread is closed whether the message is received via the same consumer or not.

All active message scopes are still closed when the session is closed.
