+++ dd-java-agent/agent-bootstrap/src/main/java/datadog/trace/bootstrap/instrumentation/jms/SessionState.java
@@ -1,10 +1,9 @@
 package datadog.trace.bootstrap.instrumentation.jms;
 
-import static java.util.Collections.newSetFromMap;
-
+import datadog.trace.bootstrap.instrumentation.api.AgentScope;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
+import java.util.Map;
 import java.util.Queue;
-import java.util.Set;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
@@ -28,8 +27,7 @@ public final class SessionState {
   private final int ackMode;
 
   // consumer-related session state
-  private final Set<MessageConsumerState> consumerStates =
-      newSetFromMap(new ConcurrentHashMap<MessageConsumerState, Boolean>());
+  private final Map<Thread, AgentScope> activeScopes = new ConcurrentHashMap<>();
   private volatile Queue<AgentSpan> capturedSpans;
   private volatile int spanCount;
 
@@ -93,20 +91,33 @@ public final class SessionState {
     }
   }
 
-  void registerConsumerState(MessageConsumerState consumerState) {
-    consumerStates.add(consumerState);
+  /** Closes the given message scope when the next message is consumed or the session is closed. */
+  void closeOnIteration(AgentScope newScope) {
+    maybeCloseScope(activeScopes.put(Thread.currentThread(), newScope));
   }
 
-  void unregisterConsumerState(MessageConsumerState consumerState) {
-    consumerStates.remove(consumerState);
+  /** Closes the scope previously registered by closeOnIteration, assumes same calling thread. */
+  void closePreviousMessageScope() {
+    maybeCloseScope(activeScopes.remove(Thread.currentThread()));
   }
 
+  /** Closes any active message scopes and finishes any pending transacted spans. */
   public void onClose() {
-    for (MessageConsumerState consumerState : consumerStates) {
-      consumerState.onClose(); // eventually calls unregisterConsumerState
+    for (AgentScope scope : activeScopes.values()) {
+      maybeCloseScope(scope);
     }
+    activeScopes.clear();
     if (isTransactedSession()) {
       onCommitOrRollback(); // implicit rollback of any active transaction
+    }
+  }
+
+  private void maybeCloseScope(AgentScope scope) {
+    if (null != scope) {
+      scope.close();
+      if (isAutoAcknowledge()) {
+        scope.span().finish();
+      }
     }
   }
 }
